test: factor out common code, and split stable tests into a separate file
